+++ client/src/components/forms/PositionForm.tsx
@@ -2,24 +2,19 @@
 
 import { BreadcrumbLink } from "@/src/components/ui/breadcrumbs/BreadcrumbLink";
 import Button from "@/src/components/ui/button";
-import Dropdown from "@/src/components/ui/dropdown";
-import Field from "@/src/components/ui/field";
-import { departments } from "@/src/data/departments";
-import { useEffect } from "react";
-import { Controller, useForm } from "react-hook-form";
-
-type Mode = "create" | "edit" | "view";
-
-export type PositionFormData = {
-    name: string;
-    department_id: string | "";
-};
+import { defaultPositionValues } from "@/src/lib/forms/position/position-form-default-values";
+import { PositionFormFields } from "@/src/lib/forms/position/position-form-field";
+import { Mode } from "@/src/types/form";
+import { FormData } from "@/src/types/forms/position-form.types";
+import { useEffect, useMemo } from "react";
+import { FormProvider, useForm } from "react-hook-form";
+import FormBuilder from "../layout/FormBuilder";
 
 type PositionFormProps = {
     mode: Mode;
     title: string;
-    initialValues?: Partial<PositionFormData>;
-    onSubmit?: (data: PositionFormData) => void;
+    initialValues?: Partial<FormData>;
+    onSubmit?: (data: FormData) => void;
     cancelPath?: string;
     submitLabel?: string;
 };
@@ -33,29 +28,21 @@ const PositionForm = ({
     submitLabel,
 }: PositionFormProps) => {
     const isReadOnly = mode === "view";
+    const defaultValues: FormData = useMemo(
+        () => defaultPositionValues(initialValues),
+        [initialValues]
+    );
 
-    const {
-        register,
-        control,
-        handleSubmit,
-        reset,
-        formState: { errors },
-    } = useForm<PositionFormData>({
-        defaultValues: {
-            name: initialValues?.name ?? "",
-            department_id: initialValues?.department_id ?? "",
-        },
+    const methods = useForm<FormData>({
+        defaultValues,
         mode: "onBlur",
     });
 
     useEffect(() => {
-        reset({
-            name: initialValues?.name ?? "",
-            department_id: initialValues?.department_id ?? "",
-        });
-    }, [initialValues, reset]);
+        methods.reset(defaultValues);
+    }, [defaultValues, methods]);
 
-    const handleFormSubmit = handleSubmit((data) => {
+    const handleFormSubmit = methods.handleSubmit((data) => {
         if (isReadOnly) return;
         onSubmit?.(data);
     });
@@ -69,61 +56,12 @@ const PositionForm = ({
                 onSubmit={handleFormSubmit}
             >
                 <div className="w-full bg-white p-4 rounded-lg shadow-md flex flex-col overflow-y-auto max-h-screen min-h-125">
-                    <div className="flex flex-col md:flex-row gap-2">
-                        <div className="w-full md:w-1/2 flex flex-col gap-y-1">
-                            <p className="text-sm">Department</p>
-                            <div className="flex flex-col">
-                                <Controller
-                                    name="department_id"
-                                    control={control}
-                                    rules={{
-                                        required: isReadOnly
-                                            ? false
-                                            : "Department is required",
-                                    }}
-                                    render={({ field }) => (
-                                        <Dropdown
-                                            value={field.value}
-                                            onChange={field.onChange}
-                                            isRounded
-                                            disabled={isReadOnly}
-                                            options={departments.map(
-                                                (department) => ({
-                                                    label: department.name,
-                                                    value: department.id,
-                                                })
-                                            )}
-                                        />
-                                    )}
-                                />
-                                {errors.department_id?.message && (
-                                    <p className="text-sm text-right pr-3 text-red-400">
-                                        {errors.department_id.message}
-                                    </p>
-                                )}
-                            </div>
-                        </div>
-                        <div className="w-full md:w-1/2 flex flex-col gap-y-1">
-                            <p className="text-sm">Name</p>
-                            <div className="flex flex-col">
-                                <Field
-                                    variant="outlined"
-                                    isRounded
-                                    disabled={isReadOnly}
-                                    {...register("name", {
-                                        required: isReadOnly
-                                            ? false
-                                            : "Name is required",
-                                    })}
-                                />
-                                {errors.name?.message && (
-                                    <p className="text-sm text-right pr-3 text-red-400">
-                                        {errors.name.message}
-                                    </p>
-                                )}
-                            </div>
-                        </div>
-                    </div>
+                    <FormProvider {...methods}>
+                        <FormBuilder
+                            fields={PositionFormFields}
+                            isReadOnly={isReadOnly}
+                        />
+                    </FormProvider>
                 </div>
                 <div className="flex flex-row items-center justify-end gap-x-3">
                     <BreadcrumbLink