+++ client/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### 🚀 Added
+- Created form builder for faster build of fields
 
 ### 🔄 Changed