+++ client/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### 🚀 Added
 - Created form builder for faster build of fields
 - Datepicker component
+- Textarea component
 
 ### 🔄 Changed