+++ client/CHANGELOG.md
@@ -14,6 +14,8 @@ All notable changes to this project will be documented in this file.
 
 ### 🔄 Changed
 - Authenticated design for the main component
+- Change department form structure and use reusable component for form builder
+- Separate files for more readability and clean file structure (Department)
 
 ### 🐛 Fixed