+++ client/CHANGELOG.md
@@ -18,6 +18,8 @@ All notable changes to this project will be documented in this file.
 - Separate files for more readability and clean file structure (Department)
 - Change position form structure and use reusable component for form builder
 - Separate files for more readability and clean file structure (Position)
+- Change service form structure and use reusable component for form builder
+- Separate files for more readability and clean file structure (Service)
 
 ### 🐛 Fixed